moving myBatis files
