+++ src/main/java/mybatis/dao/CreatureMapper.java
@@ -1,4 +1,4 @@
-package usecases.mybatis.dao;
+package mybatis.dao;
 
 import java.util.List;
 import org.apache.ibatis.annotations.Delete;
@@ -10,82 +10,80 @@ import org.apache.ibatis.annotations.Select;
 import org.apache.ibatis.annotations.Update;
 import org.apache.ibatis.type.JdbcType;
 import org.mybatis.cdi.Mapper;
-import usecases.mybatis.model.Hero;
+import mybatis.model.Creature;
+
 @Mapper
-public interface HeroMapper {
+public interface CreatureMapper {
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table PUBLIC.HERO
+     * This method corresponds to the database table PUBLIC.CREATURE
      *
      * @mbg.generated Fri Mar 08 22:23:27 EET 2019
      */
     @Delete({
-        "delete from PUBLIC.HERO",
+        "delete from PUBLIC.CREATURE",
         "where ID = #{id,jdbcType=INTEGER}"
     })
     int deleteByPrimaryKey(Integer id);
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table PUBLIC.HERO
+     * This method corresponds to the database table PUBLIC.CREATURE
      *
      * @mbg.generated Fri Mar 08 22:23:27 EET 2019
      */
     @Insert({
-        "insert into PUBLIC.HERO (FACTION, NAME)",
-        "values (#{faction,jdbcType=VARCHAR}, #{name,jdbcType=VARCHAR})"
+        "insert into PUBLIC.CREATURE (NAME)",
+        "values (#{name,jdbcType=VARCHAR})"
     })
     @Options(useGeneratedKeys=true,keyProperty="id")
-    int insert(Hero record);
+    int insert(Creature record);
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table PUBLIC.HERO
+     * This method corresponds to the database table PUBLIC.CREATURE
      *
      * @mbg.generated Fri Mar 08 22:23:27 EET 2019
      */
     @Select({
         "select",
-        "ID, FACTION, NAME",
-        "from PUBLIC.HERO",
+        "ID, NAME",
+        "from PUBLIC.CREATURE",
         "where ID = #{id,jdbcType=INTEGER}"
     })
     @Results({
         @Result(column="ID", property="id", jdbcType=JdbcType.INTEGER, id=true),
-        @Result(column="FACTION", property="faction", jdbcType=JdbcType.VARCHAR),
         @Result(column="NAME", property="name", jdbcType=JdbcType.VARCHAR)
     })
-    Hero selectByPrimaryKey(Integer id);
+    Creature selectByPrimaryKey(Integer id);
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table PUBLIC.HERO
+     * This method corresponds to the database table PUBLIC.CREATURE
      *
      * @mbg.generated Fri Mar 08 22:23:27 EET 2019
      */
     @Select({
         "select",
-        "ID, FACTION, NAME",
-        "from PUBLIC.HERO"
+        "ID, NAME",
+        "from PUBLIC.CREATURE"
     })
     @Results({
         @Result(column="ID", property="id", jdbcType=JdbcType.INTEGER, id=true),
-        @Result(column="FACTION", property="faction", jdbcType=JdbcType.VARCHAR),
         @Result(column="NAME", property="name", jdbcType=JdbcType.VARCHAR)
     })
-    List<Hero> selectAll();
+    List<Creature> selectAll();
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table PUBLIC.HERO
+     * This method corresponds to the database table PUBLIC.CREATURE
      *
      * @mbg.generated Fri Mar 08 22:23:27 EET 2019
      */
     @Update({
-        "update PUBLIC.HERO",
-        "set FACTION = #{faction,jdbcType=VARCHAR},",
-          "NAME = #{name,jdbcType=VARCHAR}",
+        "update PUBLIC.CREATURE",
+        "set NAME = #{name,jdbcType=VARCHAR}",
         "where ID = #{id,jdbcType=INTEGER}"
     })
-    int updateByPrimaryKey(Hero record);
+    int updateByPrimaryKey(Creature record);
 }
