+++ src/main/java/mybatis/dao/HeroMapper.java
@@ -1,4 +1,4 @@
-package usecases.mybatis.dao;
+package mybatis.dao;
 
 import java.util.List;
 import org.apache.ibatis.annotations.Delete;
@@ -10,87 +10,82 @@ import org.apache.ibatis.annotations.Select;
 import org.apache.ibatis.annotations.Update;
 import org.apache.ibatis.type.JdbcType;
 import org.mybatis.cdi.Mapper;
-import usecases.mybatis.model.Skill;
+import mybatis.model.Hero;
 @Mapper
-public interface SkillMapper {
+public interface HeroMapper {
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table PUBLIC.SKILL
+     * This method corresponds to the database table PUBLIC.HERO
      *
      * @mbg.generated Fri Mar 08 22:23:27 EET 2019
      */
     @Delete({
-        "delete from PUBLIC.SKILL",
+        "delete from PUBLIC.HERO",
         "where ID = #{id,jdbcType=INTEGER}"
     })
     int deleteByPrimaryKey(Integer id);
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table PUBLIC.SKILL
+     * This method corresponds to the database table PUBLIC.HERO
      *
      * @mbg.generated Fri Mar 08 22:23:27 EET 2019
      */
     @Insert({
-        "insert into PUBLIC.SKILL (DESCRIPTION, LEVELREQUIREMENT, ",
-        "NAME)",
-        "values (#{description,jdbcType=VARCHAR}, #{levelrequirement,jdbcType=TINYINT}, ",
-        "#{name,jdbcType=VARCHAR})"
+        "insert into PUBLIC.HERO (FACTION, NAME)",
+        "values (#{faction,jdbcType=VARCHAR}, #{name,jdbcType=VARCHAR})"
     })
     @Options(useGeneratedKeys=true,keyProperty="id")
-    int insert(Skill record);
+    int insert(Hero record);
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table PUBLIC.SKILL
+     * This method corresponds to the database table PUBLIC.HERO
      *
      * @mbg.generated Fri Mar 08 22:23:27 EET 2019
      */
     @Select({
         "select",
-        "ID, DESCRIPTION, LEVELREQUIREMENT, NAME",
-        "from PUBLIC.SKILL",
+        "ID, FACTION, NAME",
+        "from PUBLIC.HERO",
         "where ID = #{id,jdbcType=INTEGER}"
     })
     @Results({
         @Result(column="ID", property="id", jdbcType=JdbcType.INTEGER, id=true),
-        @Result(column="DESCRIPTION", property="description", jdbcType=JdbcType.VARCHAR),
-        @Result(column="LEVELREQUIREMENT", property="levelrequirement", jdbcType=JdbcType.TINYINT),
+        @Result(column="FACTION", property="faction", jdbcType=JdbcType.VARCHAR),
         @Result(column="NAME", property="name", jdbcType=JdbcType.VARCHAR)
     })
-    Skill selectByPrimaryKey(Integer id);
+    Hero selectByPrimaryKey(Integer id);
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table PUBLIC.SKILL
+     * This method corresponds to the database table PUBLIC.HERO
      *
      * @mbg.generated Fri Mar 08 22:23:27 EET 2019
      */
     @Select({
         "select",
-        "ID, DESCRIPTION, LEVELREQUIREMENT, NAME",
-        "from PUBLIC.SKILL"
+        "ID, FACTION, NAME",
+        "from PUBLIC.HERO"
     })
     @Results({
         @Result(column="ID", property="id", jdbcType=JdbcType.INTEGER, id=true),
-        @Result(column="DESCRIPTION", property="description", jdbcType=JdbcType.VARCHAR),
-        @Result(column="LEVELREQUIREMENT", property="levelrequirement", jdbcType=JdbcType.TINYINT),
+        @Result(column="FACTION", property="faction", jdbcType=JdbcType.VARCHAR),
         @Result(column="NAME", property="name", jdbcType=JdbcType.VARCHAR)
     })
-    List<Skill> selectAll();
+    List<Hero> selectAll();
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table PUBLIC.SKILL
+     * This method corresponds to the database table PUBLIC.HERO
      *
      * @mbg.generated Fri Mar 08 22:23:27 EET 2019
      */
     @Update({
-        "update PUBLIC.SKILL",
-        "set DESCRIPTION = #{description,jdbcType=VARCHAR},",
-          "LEVELREQUIREMENT = #{levelrequirement,jdbcType=TINYINT},",
+        "update PUBLIC.HERO",
+        "set FACTION = #{faction,jdbcType=VARCHAR},",
           "NAME = #{name,jdbcType=VARCHAR}",
         "where ID = #{id,jdbcType=INTEGER}"
     })
-    int updateByPrimaryKey(Skill record);
+    int updateByPrimaryKey(Hero record);
 }
