+++ src/main/java/mybatis/dao/SkillMapper.java
@@ -1,4 +1,4 @@
-package usecases.mybatis.dao;
+package mybatis.dao;
 
 import java.util.List;
 import org.apache.ibatis.annotations.Delete;
@@ -10,80 +10,87 @@ import org.apache.ibatis.annotations.Select;
 import org.apache.ibatis.annotations.Update;
 import org.apache.ibatis.type.JdbcType;
 import org.mybatis.cdi.Mapper;
-import usecases.mybatis.model.Creature;
-
+import mybatis.model.Skill;
 @Mapper
-public interface CreatureMapper {
+public interface SkillMapper {
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table PUBLIC.CREATURE
+     * This method corresponds to the database table PUBLIC.SKILL
      *
      * @mbg.generated Fri Mar 08 22:23:27 EET 2019
      */
     @Delete({
-        "delete from PUBLIC.CREATURE",
+        "delete from PUBLIC.SKILL",
         "where ID = #{id,jdbcType=INTEGER}"
     })
     int deleteByPrimaryKey(Integer id);
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table PUBLIC.CREATURE
+     * This method corresponds to the database table PUBLIC.SKILL
      *
      * @mbg.generated Fri Mar 08 22:23:27 EET 2019
      */
     @Insert({
-        "insert into PUBLIC.CREATURE (NAME)",
-        "values (#{name,jdbcType=VARCHAR})"
+        "insert into PUBLIC.SKILL (DESCRIPTION, LEVELREQUIREMENT, ",
+        "NAME)",
+        "values (#{description,jdbcType=VARCHAR}, #{levelrequirement,jdbcType=TINYINT}, ",
+        "#{name,jdbcType=VARCHAR})"
     })
     @Options(useGeneratedKeys=true,keyProperty="id")
-    int insert(Creature record);
+    int insert(Skill record);
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table PUBLIC.CREATURE
+     * This method corresponds to the database table PUBLIC.SKILL
      *
      * @mbg.generated Fri Mar 08 22:23:27 EET 2019
      */
     @Select({
         "select",
-        "ID, NAME",
-        "from PUBLIC.CREATURE",
+        "ID, DESCRIPTION, LEVELREQUIREMENT, NAME",
+        "from PUBLIC.SKILL",
         "where ID = #{id,jdbcType=INTEGER}"
     })
     @Results({
         @Result(column="ID", property="id", jdbcType=JdbcType.INTEGER, id=true),
+        @Result(column="DESCRIPTION", property="description", jdbcType=JdbcType.VARCHAR),
+        @Result(column="LEVELREQUIREMENT", property="levelrequirement", jdbcType=JdbcType.TINYINT),
         @Result(column="NAME", property="name", jdbcType=JdbcType.VARCHAR)
     })
-    Creature selectByPrimaryKey(Integer id);
+    Skill selectByPrimaryKey(Integer id);
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table PUBLIC.CREATURE
+     * This method corresponds to the database table PUBLIC.SKILL
      *
      * @mbg.generated Fri Mar 08 22:23:27 EET 2019
      */
     @Select({
         "select",
-        "ID, NAME",
-        "from PUBLIC.CREATURE"
+        "ID, DESCRIPTION, LEVELREQUIREMENT, NAME",
+        "from PUBLIC.SKILL"
     })
     @Results({
         @Result(column="ID", property="id", jdbcType=JdbcType.INTEGER, id=true),
+        @Result(column="DESCRIPTION", property="description", jdbcType=JdbcType.VARCHAR),
+        @Result(column="LEVELREQUIREMENT", property="levelrequirement", jdbcType=JdbcType.TINYINT),
         @Result(column="NAME", property="name", jdbcType=JdbcType.VARCHAR)
     })
-    List<Creature> selectAll();
+    List<Skill> selectAll();
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table PUBLIC.CREATURE
+     * This method corresponds to the database table PUBLIC.SKILL
      *
      * @mbg.generated Fri Mar 08 22:23:27 EET 2019
      */
     @Update({
-        "update PUBLIC.CREATURE",
-        "set NAME = #{name,jdbcType=VARCHAR}",
+        "update PUBLIC.SKILL",
+        "set DESCRIPTION = #{description,jdbcType=VARCHAR},",
+          "LEVELREQUIREMENT = #{levelrequirement,jdbcType=TINYINT},",
+          "NAME = #{name,jdbcType=VARCHAR}",
         "where ID = #{id,jdbcType=INTEGER}"
     })
-    int updateByPrimaryKey(Creature record);
+    int updateByPrimaryKey(Skill record);
 }
