+++ src/main/java/mybatis/dao/StatMapper.java
@@ -1,4 +1,4 @@
-package usecases.mybatis.dao;
+package mybatis.dao;
 
 import java.util.List;
 import org.apache.ibatis.annotations.Delete;
@@ -10,7 +10,7 @@ import org.apache.ibatis.annotations.Select;
 import org.apache.ibatis.annotations.Update;
 import org.apache.ibatis.type.JdbcType;
 import org.mybatis.cdi.Mapper;
-import usecases.mybatis.model.Stat;
+import mybatis.model.Stat;
 @Mapper
 public interface StatMapper {
     /**
