+++ src/main/java/usecases/MyBatisResources.java
@@ -1,4 +1,4 @@
-package usecases.cdi;
+package usecases;
 
 import org.apache.ibatis.io.Resources;
 import org.apache.ibatis.session.SqlSessionFactory;
