+++ src/main/java/usecases/cdi/jpa/Creatures.java
@@ -11,7 +11,6 @@ import javax.enterprise.inject.Model;
 import javax.inject.Inject;
 import javax.transaction.Transactional;
 import java.io.Serializable;
-import java.util.ArrayList;
 import java.util.List;
 
 @Model
