+++ src/main/resources/generatorConfig.xml
@@ -11,14 +11,14 @@
                     password="as">
     </jdbcConnection>
 
-    <javaModelGenerator targetPackage="usecases.mybatis.model" targetProject="MAVEN" />
+    <javaModelGenerator targetPackage="mybatis.model" targetProject="MAVEN" />
 
     <!-- Anotacijų variantas - mažiau galimybių, bet paprastesnis -->
-    <javaClientGenerator type="ANNOTATEDMAPPER" targetPackage="usecases.mybatis.dao" targetProject="MAVEN" />
+    <javaClientGenerator type="ANNOTATEDMAPPER" targetPackage="mybatis.dao" targetProject="MAVEN" />
 
     <!-- XML variantas - daugiau galimybių -->
-    <!--<sqlMapGenerator targetPackage=usecases.mybatis.xml" targetProject="MAVEN" />-->
-    <!--<javaClientGenerator type="XMLMAPPER" targetPackage="usecases.mybatis.dao" targetProject="MAVEN" />-->
+    <!--<sqlMapGenerator targetPackage=mybatis.xml" targetProject="MAVEN" />-->
+    <!--<javaClientGenerator type="XMLMAPPER" targetPackage="mybatis.dao" targetProject="MAVEN" />-->
 
 
     <table schema="public" tableName="creature">
